+++ demo/api/tests/integration/test_postgres_interrupt_runtime.py
@@ -10,7 +10,7 @@ import pytest
 from fastapi import FastAPI
 from httpx import ASGITransport, AsyncClient
 from langgraph_openai_serve import GraphRegistry, LanggraphOpenaiServe
-from langgraph_openai_serve.graph.utils import checkpoint_key
+from langgraph_openai_serve.graph.interrupt.state import checkpoint_key
 from openai import AsyncOpenAI, ConflictError
 
 from lgos_demo_api.checkpointer import (
