+++ src/langgraph_openai_serve/graph/utils.py
@@ -1,10 +1,7 @@
 """Prepare one isolated LangGraph execution for the OpenAI API."""
 
-import hashlib
-import json
 import logging
 import sys
-import uuid
 from contextlib import AbstractAsyncContextManager
 from dataclasses import dataclass, field
 from typing import Any, cast
@@ -16,35 +13,23 @@ from langchain_core.callbacks.base import (
     Callbacks,
 )
 from langchain_core.runnables import RunnableConfig
-from langgraph.checkpoint.base import (
-    RESUME,
-    BaseCheckpointSaver,
-    get_checkpoint_id,
-)
 from langgraph.graph.state import CompiledStateGraph
-from langgraph.types import Command, Interrupt, StateSnapshot
 
 from langgraph_openai_serve.api.chat.schemas import (
     ChatCompletionRequest,
     ChatCompletionRequestMessage,
 )
-from langgraph_openai_serve.api.chat.utils.interrupts import (
-    InterruptResume,
-    InvalidResumeRequestError,
-    parse_resume_request,
-)
+from langgraph_openai_serve.api.chat.utils.interrupts import parse_resume_request
 from langgraph_openai_serve.core.settings import settings
 from langgraph_openai_serve.graph.features import GraphFeature
 from langgraph_openai_serve.graph.graph_registry import (
     GraphConfig,
-    GraphConfigurationError,
     GraphRegistry,
 )
+from langgraph_openai_serve.graph.interrupt import state as interrupt_state
 from langgraph_openai_serve.utils.message import convert_to_lc_messages
 
 logger = logging.getLogger(__name__)
-
-RUN_METADATA_KEY = "langgraph_run_id"
 
 if settings.ENABLE_LANGFUSE is True:
     from langfuse.langchain import CallbackHandler
@@ -72,14 +57,6 @@ class GraphRun:
         lease, self._lease = self._lease, None
         if lease is not None:
             await lease.__aexit__(None, None, None)
-
-
-class InvalidRunIDError(ValueError):
-    """Raised when a caller-supplied run id is not a UUID."""
-
-
-class InterruptStateConflictError(RuntimeError):
-    """Raised when a resume does not match durable pending state."""
 
 
 async def prepare_run(
@@ -111,12 +88,12 @@ async def prepare_run(
         )
 
     resume = parse_resume_request(messages)
-    requested_run_id = get_run_id(request)
-    run_id = _resolve_run_id(requested_run_id, resume)
-    checkpoint_thread_id = checkpoint_key(
+    requested_run_id = interrupt_state.get_run_id(request)
+    run_id = interrupt_state.resolve_run_id(requested_run_id, resume)
+    checkpoint_thread_id = interrupt_state.checkpoint_key(
         model,
         run_id,
-        scope=normalize_checkpoint_scope(checkpoint_scope),
+        scope=interrupt_state.normalize_checkpoint_scope(checkpoint_scope),
     )
     runnable_config = build_runnable_config(
         graph_config.runtime_callbacks,
@@ -133,7 +110,7 @@ async def prepare_run(
 
     try:
         snapshot = await graph.aget_state(runnable_config, subgraphs=True)
-        inputs, should_execute = await _prepare_interrupt_input(
+        inputs, should_execute = await interrupt_state.prepare_interrupt_input(
             graph_config,
             graph,
             request,
@@ -165,191 +142,6 @@ async def prepare_run(
         should_execute=should_execute,
         _lease=lease,
     )
-
-
-async def _prepare_interrupt_input(
-    graph_config: GraphConfig,
-    graph: CompiledStateGraph,
-    request: ChatCompletionRequest,
-    snapshot: StateSnapshot,
-    resume: InterruptResume | None,
-) -> tuple[Any, bool]:
-    pending_interrupts = _interrupts_by_id(snapshot)
-    checkpoint_id = get_checkpoint_id(snapshot.config)
-
-    if resume is None:
-        if checkpoint_id is None:
-            lc_messages = convert_to_lc_messages(request.messages)
-            return await graph_config.build_input(request, lc_messages), True
-        if pending_interrupts:
-            # Re-emit persisted tool calls without rerunning graph nodes.
-            return None, False
-        raise InterruptStateConflictError("This run_id has already been used.")
-
-    if checkpoint_id is None:
-        raise InterruptStateConflictError(
-            "No durable interrupt state exists for this run."
-        )
-    if not pending_interrupts:
-        raise InterruptStateConflictError("This run no longer has pending interrupts.")
-
-    state_token = await checkpoint_state_token(graph, snapshot.config)
-    if state_token is None:
-        raise InterruptStateConflictError(
-            "No durable interrupt state exists for this run."
-        )
-    return (
-        _resume_interrupt_inputs(state_token, set(pending_interrupts), resume),
-        True,
-    )
-
-
-def _resume_interrupt_inputs(
-    state_token: str,
-    pending_ids: set[str],
-    resume: InterruptResume,
-) -> Command:
-    if resume.state_token != state_token:
-        raise InterruptStateConflictError(
-            "The interrupt result is stale for the current interrupt generation."
-        )
-    if set(resume.values) != pending_ids:
-        raise InterruptStateConflictError(
-            "Interrupt results do not match the complete pending interrupt set."
-        )
-
-    # Always use the ID/value form, including for one interrupt. It preserves
-    # OpenAI tool_call causality and handles JSON null as a legitimate answer.
-    return Command(resume=resume.values)
-
-
-def _interrupts_by_id(snapshot: StateSnapshot) -> dict[str, Interrupt]:
-    pending: dict[str, Interrupt] = {}
-    for interrupt in snapshot.interrupts:
-        interrupt_id = interrupt.id
-        if not isinstance(interrupt_id, str) or not interrupt_id:
-            raise RuntimeError("Durable interrupt state has an invalid interrupt id.")
-        if interrupt_id in pending:
-            raise RuntimeError("Durable interrupt state has duplicate interrupt ids.")
-        pending[interrupt_id] = interrupt
-    return pending
-
-
-def _resolve_run_id(
-    requested_run_id: str | None,
-    resume: InterruptResume | None,
-) -> str:
-    if resume is not None:
-        resume_run_id = normalize_run_id(resume.run_id)
-        if requested_run_id is not None:
-            requested_run_id = normalize_run_id(requested_run_id)
-            if requested_run_id != resume_run_id:
-                raise InvalidResumeRequestError(
-                    f"metadata.{RUN_METADATA_KEY} does not match the interrupt tool call."
-                )
-        return resume_run_id
-
-    if requested_run_id is not None:
-        return normalize_run_id(requested_run_id)
-    return str(uuid.uuid4())
-
-
-def normalize_run_id(value: str) -> str:
-    try:
-        parsed = uuid.UUID(value)
-    except (AttributeError, TypeError, ValueError) as exc:
-        raise InvalidRunIDError(
-            f"metadata.{RUN_METADATA_KEY} must be a UUID when provided."
-        ) from exc
-    if parsed.int == 0:
-        raise InvalidRunIDError(
-            f"metadata.{RUN_METADATA_KEY} must not be the nil UUID."
-        )
-    return str(parsed)
-
-
-def get_run_id(request: ChatCompletionRequest) -> str | None:
-    return (request.metadata or {}).get(RUN_METADATA_KEY)
-
-
-def normalize_checkpoint_scope(value: str) -> str:
-    if not isinstance(value, str) or not value.strip():
-        raise GraphConfigurationError(
-            "checkpoint_scope must resolve to a non-empty server-trusted string."
-        )
-    return value.strip()
-
-
-def checkpoint_key(model: str, run_id: str, *, scope: str = "default") -> str:
-    """Derive a fixed-length storage key scoped to this protocol and model."""
-    identity = json.dumps(
-        ["langgraph-openai-serve.interrupt.v2", scope, model, run_id],
-        ensure_ascii=False,
-        separators=(",", ":"),
-    )
-    return hashlib.sha256(identity.encode()).hexdigest()
-
-
-async def checkpoint_state_token(
-    graph: CompiledStateGraph,
-    runnable_config: RunnableConfig,
-) -> str | None:
-    """Fingerprint the latest checkpoint in every namespace.
-
-    Nested resumes may not advance the root checkpoint, and indirectly invoked
-    subgraphs are not exposed through state snapshots. Scanning the checkpointer
-    keeps stale-resume detection generic without introducing separate state.
-
-    Performance impact: Local PostgreSQL measurements were 0.5-0.7 ms for the current 1-2 tuple
-    runs, scaling linearly to about 5 ms at 100 and 45 ms at 1,000 small tuples.
-    """
-    checkpointer = cast(BaseCheckpointSaver, graph.checkpointer)
-    thread_id = runnable_config["configurable"]["thread_id"]
-    heads: dict[str, tuple[str, list[tuple[str, int]]]] = {}
-
-    async for checkpoint_tuple in checkpointer.alist(
-        {"configurable": {"thread_id": thread_id}}
-    ):
-        namespace = checkpoint_tuple.config["configurable"].get("checkpoint_ns", "")
-        checkpoint_id = require_checkpoint_id(checkpoint_tuple.config)
-        head = heads.get(namespace)
-        if head is not None and checkpoint_id <= head[0]:
-            continue
-
-        heads[namespace] = (
-            checkpoint_id,
-            sorted(
-                (
-                    task_id,
-                    len(value) if isinstance(value, (list, tuple)) else 1,
-                )
-                for task_id, channel, value in checkpoint_tuple.pending_writes or ()
-                if channel == RESUME
-            ),
-        )
-
-    if not heads:
-        return None
-
-    identity = json.dumps(
-        [
-            "langgraph-openai-serve.interrupt-state.v2",
-            sorted((namespace, *head) for namespace, head in heads.items()),
-        ],
-        separators=(",", ":"),
-    )
-    return hashlib.sha256(identity.encode()).hexdigest()
-
-
-def require_checkpoint_id(config: RunnableConfig) -> str:
-    """Return the checkpoint id from a validated LangGraph config."""
-    try:
-        checkpoint_id = get_checkpoint_id(config)
-    except (AttributeError, KeyError, TypeError):
-        checkpoint_id = None
-    if not isinstance(checkpoint_id, str) or not checkpoint_id:
-        raise RuntimeError("Durable interrupt state has no checkpoint_id.")
-    return checkpoint_id
 
 
 def build_runnable_config(
