+++ tests/api/interrupt/conftest.py
@@ -13,7 +13,7 @@ from langgraph_openai_serve import (
     GraphRegistry,
     LanggraphOpenaiServe,
 )
-from langgraph_openai_serve.graph.coordination import InMemoryRunCoordinator
+from langgraph_openai_serve.graph.interrupt import InMemoryRunCoordinator
 from tests.graph.support.interrupt import (
     InterruptAnswerState,
     make_interrupt_graph,
