+++ tests/graph/interrupt/__init__.py
@@ -1,0 +1,1 @@
+"""Graph interrupt test package."""
